+++ modules/canvaskit/canvaskit_bindings.cpp
@@ -1767,6 +1767,8 @@ EMSCRIPTEN_BINDINGS(Skia) {
         .field("columns", &RuntimeEffectUniform::columns)
         .field("rows",    &RuntimeEffectUniform::rows)
         .field("slot",    &RuntimeEffectUniform::slot);
+
+    constant("rt_effect", true);
 #endif
 
     class_<SkSurface>("Surface")