+++ modules/canvaskit/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - A full build of CanvasKit is now in /bin/full.
+
 ## [0.24.0] - 2021-02-18
 
 ### Added